requirements.txt
numpy
pandas
scikit-learn
xgboost
pytorch-tabnet
torch
torchvision
matplotlib
seaborn
joblib
pyyaml
